+++ app/models/user.ts
@@ -35,6 +35,9 @@ export default class User extends OsfModel.extend(Validations) {
     @attr('array') middleNames!: string[];
     @attr('fixstring') familyName!: string;
 
+    @attr('fixstring') locale!: string;
+    @attr('fixstring') timezone!: string;
+
     @attr('date') dateRegistered!: Date;
     // email
     @attr('fixstring') username!: string;