+++ types/ember-metrics/index.d.ts
@@ -3,6 +3,10 @@ declare module 'ember-metrics/services/metrics' {
 
     export default class Metrics extends Service {
         trackEvent(params: object): void;
+
+        /* eslint-disable no-dupe-class-members */
         trackPage(params: object): void;
+        trackPage(adapterName: string, params: object): void;
+        /* eslint-enable no-dupe-class-members */
     }
 }